DAY7: Some math problems
I'd have to take a little more time to work out the math on this
problem. I am not positive this is actually correct. We could have
done an expensive calculation iterating through all of the possible
starting points and seeing what the fuel cost is and just keeping
the smallest, but the math solution seems nicer if it holds.