DAY8: Logic puzzle
This one took quite a bit of reading to understand what was being
asked. After understanding the problem, part 1 was pretty straight
forward. Part 2 on the other hand, seemed like I had to figure out
how to solve the problem and then try to code that. I'm not sure the
code worked out too well for me, seemed a bit all over the place, but
that is partly due to how I solved the problem in my head by back
calculating the pieces.