DAY11: updating neighbors with a stack
Used a stack to process new elements and recheck only the
necessary items. I made a mistake at first by replicating a
mutable list to keep track of the flashes. [[False]*10]*10 shares
the pointer to the rows, which meant that I was updating more
flashes than I thought at first. Took a bit of time to debug that
mistake, but obvious once I found it and hopefully will learn
from that for the future!