DAY12: path traversals
Messed around with a few different data structures going through this
and what to pass into the recursive path traversal function and
what to return from it. The second problem is still pretty slow,
so there is definitely something we should improve upon there. We
could probably speed up what checks are being done for the "two"
small tasks because we do a n2 search for that. I tried to change
the visited list to a visited dictionary instead, but that complicated
things by squashing duplicate paths into the same one because the keys
were no longer unique for each path. So we might have needed to pass
that separate dictionary along as well instead of computing it every
time.