DAY15: Maze search problem
I made a quick to implement recursive search and added a few small
optimizations for ending the search space a little quicker if a
faster path had already been found at that node.

It would be good to come back to this problem and try to code up
A* or some other thing without looking at the solutions.