DAY16: Packet decomposition
This one was a bit tricky to decipher what was going on in the
multi-packet wording. I started out trying to pass in reduced
bit strings, but then ended up going to just keeping track of the
position which seemed to work better in this case to not mess with
the lengths of bit strings and have to recompose things later.

Part 2 went really fast after the first part was decomposed properly.