DAY18: Binary tree, or list manipulation
I started off thinking of this like a binary tree with left/right
nodes with values that we could do traversals of. I think that is
actually a more elegant solution, but would have required a little
more work with the find_left / find_right methods. Instead, I pivoted
to just do in-place list replacements, which was actually not as
hard to implement as I expected. It is definitely slow to insert
into the middle of lists, but luckily this problem didn't have
long lists to deal with.