DAY19: Mathematical edges
This was the toughest one yet for me. I started off thinking too
mathematically and trying to find the optimal solution. Which would
involve solving for edges being the same (beacon1 -> beacon2) as the
subtraction is invariant between the systems as long as the coordinate
system was found first. However, this was more time consuming to
calculate the edges and compare differences and I moved back to making
a scanner class that could be rotated/moved and then just moved it to
each of the locations shared between two scanners. Brute force approach
but it did the job. This would be another good problem to come back
to and try to figure out a more optimal way of solving for uniqueness
here, potentially only using a few points in each scanner rather than
iterating through large arrays with multiple nests like I did.