DAY20: Grid expansion
A fun image enhancement problem. It took a bit of thinking about
how to handle the all on/off towards the infinities. There is probably
a mapping that could be kept track of as well, but this felt like a
satisfactory way of doing it by swapping only the pixels we needed
at the boundaries and growing the grid each time.