DAY21: Dice rolling
Easy first part, then a bit of thinking to get the second part, but
turned out to not be super difficult either. I implemented a quick
recursive implementation going in the possible paths, and then
multiplying by the "weight" (number of times that path was chosen)
when returning the number of outcomes.