DAY22: Cube addition/subtraction in large domain
I saw the expansion coming, so didn't even bother with the
naive approach. My solution involved having on/off cubes
in lists to keep track of the state. This involved a small
intersection routine to create new cubes for the intersections
that were used to add/subtract from the total volume.