DAY23: Recursive maze-like problem
Implemented this with a copy of the entire problem when recursing,
but that took a little bit of time. Pivoted to "undoing" the move
on the way out of the recursive function, which made things much faster.

I also didn't bother making a proper "reader" for this one and just
hard-coded in the game board as it was small enough.