DAY24: Long digit recursive search
I saw that we had 14 digits to traverse which would be too many to
tackle with a direct search. I decided to go backwards instead.
Realizing that we knew which cases made it to the end if we started
at the end instead. To make sure we didn't do a direct recursive search
I decided to keep track of all good input z values and then pass those
into the recursive search, so we could check many more possibilities
at once without having to redo certain combinations.