DAY25: Moving cucumbers on the map
Finished off with a bit of an easier puzzle. I ended up going with
a list of elements, but rather than iterating through each time,
making sure we only moved the pieces that were able to move.

Had fun doing these puzzles and thinking about data structures and
what to use for a given problem.